+++ examples/flying_dual_arm_ur5e.py
@@ -3,7 +3,6 @@ from pathlib import Path
 import mujoco
 import mujoco.viewer
 import numpy as np
-from dm_control import mjcf
 from loop_rate_limiters import RateLimiter
 
 import mink
@@ -12,75 +11,66 @@ _HERE = Path(__file__).parent
 _XML = _HERE / "universal_robots_ur5e" / "ur5e.xml"
 
 
-def construct_model():
-    root = mjcf.RootElement()
-    root.statistic.meansize = 0.08
-    getattr(root.visual, "global").azimuth = -120
-    getattr(root.visual, "global").elevation = -20
+def construct_model() -> mujoco.MjModel:
+    root = mujoco.MjSpec()
+    root.stat.meansize = 0.08
+    root.visual.global_.azimuth = -120
+    root.visual.global_.elevation = -20
 
-    root.worldbody.add("light", pos="0 0 1.5", directional="true")
+    root.worldbody.add_light(pos=(0, 0, 1.5), directional=True)
 
-    base = root.worldbody.add("body", name="base")
+    base = root.worldbody.add_body(name="base")
     width, height, depth = 0.8, 0.4, 0.2
-    base.add(
-        "geom",
-        type="box",
+    base.add_geom(
+        type=mujoco.mjtGeom.mjGEOM_BOX,
         size=[width, height, depth],
         density=1e-3,
-        rgba=".9 .8 .6 1",
+        rgba=(0.9, 0.8, 0.6, 1),
     )
     base.pos = [-0.0 * width, -0.0 * height, -0.5 * depth]
-    base.add("freejoint")
-    base.add("site", name="base", pos=[0, 0, depth], group=1)
-    left_site = base.add("site", name="l_attachment_site", pos=[0.3, 0, depth], group=5)
-    right_site = base.add(
-        "site",
-        name="r_attachment_site",
-        pos=[-0.3, 0, depth],
-        group=5,
+    base.add_freejoint()
+    base.add_site(name="base", pos=(0, 0, depth), group=1)
+    left_site = base.add_site(name="l_attachment_site", pos=(0.3, 0, depth), group=5)
+    right_site = base.add_site(name="r_attachment_site", pos=(-0.3, 0, depth), group=5)
+
+    left_ur5e = mujoco.MjSpec.from_file(_XML.as_posix())
+    left_ur5e.modelname = "l_ur5e"
+    left_ur5e.key("home").delete()
+    root.attach(left_ur5e, prefix="l_ur5e/", site=left_site)
+
+    right_ur5e = mujoco.MjSpec.from_file(_XML.as_posix())
+    right_ur5e.modelname = "r_ur5e"
+    right_ur5e.key("home").delete()
+    root.attach(right_ur5e, prefix="r_ur5e/", site=right_site)
+
+    body = root.worldbody.add_body(name="base_target", mocap=True)
+    body.add_geom(
+        type=mujoco.mjtGeom.mjGEOM_BOX,
+        size=(0.05, 0.05, 0.05),
+        contype=0,
+        conaffinity=0,
+        rgba=(0.6, 0.3, 0.3, 0.5),
     )
 
-    left_ur5e = mjcf.from_path(_XML.as_posix())
-    left_ur5e.model = "l_ur5e"
-    left_ur5e.find("key", "home").remove()
-    left_site.attach(left_ur5e)
-
-    right_ur5e = mjcf.from_path(_XML.as_posix())
-    right_ur5e.model = "r_ur5e"
-    right_ur5e.find("key", "home").remove()
-    right_site.attach(right_ur5e)
-
-    body = root.worldbody.add("body", name="base_target", mocap=True)
-    body.add(
-        "geom",
-        type="box",
-        size=".05 .05 .05",
-        contype="0",
-        conaffinity="0",
-        rgba=".6 .3 .3 .5",
+    body = root.worldbody.add_body(name="l_target", mocap=True)
+    body.add_geom(
+        type=mujoco.mjtGeom.mjGEOM_BOX,
+        size=(0.05, 0.05, 0.05),
+        contype=0,
+        conaffinity=0,
+        rgba=(0.3, 0.6, 0.3, 0.5),
     )
 
-    body = root.worldbody.add("body", name="l_target", mocap=True)
-    body.add(
-        "geom",
-        type="box",
-        size=".05 .05 .05",
-        contype="0",
-        conaffinity="0",
-        rgba=".3 .6 .3 .5",
+    body = root.worldbody.add_body(name="r_target", mocap=True)
+    body.add_geom(
+        type=mujoco.mjtGeom.mjGEOM_BOX,
+        size=(0.05, 0.05, 0.05),
+        contype=0,
+        conaffinity=0,
+        rgba=(0.3, 0.3, 0.6, 0.5),
     )
 
-    body = root.worldbody.add("body", name="r_target", mocap=True)
-    body.add(
-        "geom",
-        type="box",
-        size=".05 .05 .05",
-        contype="0",
-        conaffinity="0",
-        rgba=".3 .3 .6 .5",
-    )
-
-    return mujoco.MjModel.from_xml_string(root.to_xml_string(), root.get_assets())
+    return root.compile()
 
 
 if __name__ == "__main__":